# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NaikPete' API",
        "description": "API documentation for My Application",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/search-location": {
            "get": {
                "tags": [
                    "Geocoding"
                ],
                "summary": "Search location by query",
                "operationId": "3936f4cb0f623de7538fa607c296c379",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Location to search for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Universitas Ciputra"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "place_name": {
                                                "type": "string"
                                            },
                                            "lat": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "lon": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch location",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch location"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reverse-geocode": {
            "get": {
                "tags": [
                    "Geocoding"
                ],
                "summary": "Get address by latitude and longitude",
                "operationId": "8b88d52b06d01d3e42203acbe3ce59c0",
                "parameters": [
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Latitude of the location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": "-5.2299394697095245"
                        }
                    },
                    {
                        "name": "lon",
                        "in": "query",
                        "description": "Longitude of the location",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": "119.50211253693419"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address for given coordinates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "address": {
                                            "type": "object"
                                        },
                                        "place_name": {
                                            "type": "string"
                                        },
                                        "lat": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "lon": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch address",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get-route": {
            "get": {
                "tags": [
                    "Geocoding"
                ],
                "summary": "Get route from start to end location",
                "operationId": "d650ee196c6601f10128deeb83472da3",
                "parameters": [
                    {
                        "name": "start_lat",
                        "in": "query",
                        "description": "Starting latitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "start_lon",
                        "in": "query",
                        "description": "Starting longitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "end_lat",
                        "in": "query",
                        "description": "Ending latitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "end_lon",
                        "in": "query",
                        "description": "Ending longitude",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Route details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "routes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "distance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1500
                                                    },
                                                    "duration": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10
                                                    },
                                                    "geometry": {
                                                        "type": "string",
                                                        "example": "polyline_string_here"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing required parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch route from OSRM",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch route from OSRM"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/get-circular-route": {
            "get": {
                "tags": [
                    "Geocoding"
                ],
                "summary": "Get circular route by waypoints",
                "operationId": "38be7d8e9305a4769e9cdc486a5ec9a2",
                "parameters": [
                    {
                        "name": "waypoints",
                        "in": "query",
                        "description": "Waypoints in lon,lat format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "119.50211253693419,-5.2299394697095245;119.46426590725011,-5.18092218503825"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Circular route details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "routes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "distance": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1500
                                                    },
                                                    "duration": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10
                                                    },
                                                    "geometry": {
                                                        "type": "string",
                                                        "example": "polyline_string_here"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing waypoints parameter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing required parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch route from OSRM",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Failed to fetch route from OSRM"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}